requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pymysql
python-jose[cryptography]
passlib[bcrypt]
pymongo
python-dotenv
